+++ runtime/li_rt_lig.c
@@ -9,4 +9,7 @@ int32_t li_rt_lig_kernel_run(int32_t kid, int32_t bid) {
   return 0;
 }
 
-float li_rt_lig_kernel_last_validity_ratio(void) { return g_ratio; }
+/* Li `float` is 64-bit (float64), so externs declared `-> float` must return
+ * double here: a C `float` return would leave the f32 bits in the low half of
+ * the f64 register (e.g. 1.0f read back as a 5.26e-315 denormal). */
+double li_rt_lig_kernel_last_validity_ratio(void) { return (double)g_ratio; }
